pet shop solution arquitected
Proposed architecture for a cloud-based solution for a pet store.
AWS EFS was activated, that is, a common folder was created for all business instances (3 or more stores), where the 3 stores upload their files to a local repository in their availability zone and from each of these repositories they are elevated to a central repository (EFS) provided with an access protection system (Security Group). Each of the server terminals will upload and download data to and from the central repository.